Use ssize_t instead of int.
git-svn-id: svn+ssh://svn.gnucash.org/repo/gnucash/branches/gnucash-gnome2-dev@11227 57a11ea4-9604-0410-9ed3-97b8803252fd

+++ src/import-export/hbci/gnc-hbci-utils.c
@@ -904,7 +904,7 @@ gchar *gnc_call_iconv(iconv_t handler, const char* input)
 {
   char *inbuffer = (char*)input;
   char *outbuffer, *outbufferstart;
-  int inbytes, outbytes;
+  size_t inbytes, outbytes;
 
   inbytes = strlen(inbuffer);
   outbytes = inbytes + 2;
